refac: entity 합성 로직 model/index에 배치
