+++ src/basic/models/cart.ts
@@ -1,43 +1,4 @@
-import { CartItem, Product, Coupon } from "../types";
-import * as discount from "./discount";
-
-
-// 장바구니 전체 총액 계산 (할인 전/후 + 쿠폰 적용)
-export const calculateCartTotal = (
-  cart: CartItem[],
-  selectedCoupon?: Coupon | null
-): {
-  totalBeforeDiscount: number;
-  totalAfterDiscount: number;
-} => {
-  let totalBeforeDiscount = 0;
-  let totalAfterDiscount = 0;
-
-  cart.forEach((item) => {
-    const itemPrice = item.product.price * item.quantity;
-    totalBeforeDiscount += itemPrice;
-    totalAfterDiscount += discount.calculateItemTotal(item, cart);
-  });
-
-  if (selectedCoupon) {
-    if (selectedCoupon.discountType === "amount") {
-      totalAfterDiscount = Math.max(
-        0,
-        totalAfterDiscount - selectedCoupon.discountValue
-      );
-    } else {
-      totalAfterDiscount = Math.round(
-        totalAfterDiscount * (1 - selectedCoupon.discountValue / 100)
-      );
-    }
-  }
-
-  return {
-    totalBeforeDiscount: Math.round(totalBeforeDiscount),
-    totalAfterDiscount: Math.round(totalAfterDiscount),
-  };
-};
-
+import { CartItem, Product } from "../types";
 
 // 장바구니에 상품 추가
 export const addItemToCart = (
